Python3: int() conversion unnecessary in toolpalette due to use of // operator

+++ qtwidgets/toolpalette.py
@@ -297,7 +297,7 @@ class ToolPalette(QScrollArea):
         # How many widgets can fit in a row?
         # TODO: Work out hwy I need layout_spacing*3 here (we are getting the width of the scroll area, 
         # so need to take into account the borders around the grid layout? What are they?)
-        num_widgets_per_row = int((layout_width-layout_spacing*3)//(max_width+layout_spacing))
+        num_widgets_per_row = (layout_width-layout_spacing*3)//(max_width+layout_spacing)
 
         # print self._name
         # print 'number_of_widgets: %d'%len(self._widget_list)
